fix: resolve v8go BigInt crash, OOM, and test failures
- Use int32 instead of int64 for NewValue() calls that are read back
  with .Integer() — int64/uint64 create V8 BigInt, not Number, causing
  FromJust SIGTRAP on .Integer()
- Check timedOut flag in fetchFn.Call() error path, not just panic
  recovery — TerminateExecution returns an error, not a Go panic
- Apply WithResourceConstraints to V8 isolates using MemoryLimitMB
  config to prevent unbounded memory growth (was 3.7GB OOM)
- Fix timer test to use realistic wall-clock delays (10ms interval,
  50ms waits) matching the 10ms minimum interval enforcement
- Fix NoDefaultExport test: remove export keyword since V8 scripts
  don't support ES module syntax

+++ internal/worker/crypto_ext.go
@@ -170,7 +170,7 @@ func setupCryptoExt(iso *v8.Isolate, ctx *v8.Context, el *eventLoop) error {
 				namedCurve: namedCurve,
 				ecKey:      pubKey,
 			})
-			val, _ := v8.NewValue(iso, id)
+			val, _ := v8.NewValue(iso, int32(id))
 			return val
 		}
 
@@ -178,7 +178,7 @@ func setupCryptoExt(iso *v8.Isolate, ctx *v8.Context, el *eventLoop) error {
 		if id < 0 {
 			return throwError(iso, "importKey: no active request state")
 		}
-		val, _ := v8.NewValue(iso, id)
+		val, _ := v8.NewValue(iso, int32(id))
 		return val
 	}).GetFunction(ctx))
